# Conflicts:
#	components/operations.js
#	config/configUtils.js
#	utility/common_utils.js

+++ components/operations.js
@@ -6,7 +6,6 @@ const path = require('path');
 const tar = require('tar-fs');
 const uuidV4 = require('uuid').v4;
 const normalize = require('normalize-path');
-const { parentPort } = require('worker_threads');
 
 const validator = require('./operationsValidation');
 const log = require('../utility/logging/harper_logger');
@@ -42,7 +41,6 @@ function isCFEnabled() {
  * @return Object.<String>
  */
 function customFunctionsStatus() {
-	console.log(`getting custom api status`);
 	log.trace(`getting custom api status`);
 	let response = {};
 
@@ -336,25 +334,20 @@ async function packageComponent(req) {
 		}
 	}
 
-	// npm requires the contents of a module to be in a 'package' directory.
-	const tmp_project_dir = path.join(TMP_PATH, project);
-	const tmp_package_dir = path.join(tmp_project_dir, 'package');
-	fs.ensureDirSync(tmp_package_dir);
-	await fs.copy(path_to_project, tmp_package_dir, { overwrite: true });
-
+	await fs.ensureDir(TMP_PATH);
 	const file = path.join(TMP_PATH, `${project}.tar`);
 	let tar_opts = {};
 	if (req.skip_node_modules === true || req.skip_node_modules === 'true') {
 		// Create options for tar module that will exclude the CF projects node_modules directory.
 		tar_opts = {
 			ignore: (name) => {
-				return name.includes(path.join(tmp_package_dir, 'node_modules'));
+				return name.includes(path.join(path_to_project, 'node_modules'));
 			},
 		};
 	}
 
 	// pack the directory
-	tar.pack(tmp_project_dir, tar_opts).pipe(fs.createWriteStream(file, { overwrite: true }));
+	tar.pack(path_to_project, tar_opts).pipe(fs.createWriteStream(file, { overwrite: true }));
 
 	// wait for a second
 	// eslint-disable-next-line no-magic-numbers
@@ -363,7 +356,6 @@ async function packageComponent(req) {
 	// read the output into base64
 	const payload = fs.readFileSync(file, { encoding: 'base64' });
 
-	await fs.remove(tmp_project_dir);
 	await fs.remove(file);
 
 	// return the package payload as base64-encoded string
@@ -373,7 +365,7 @@ async function packageComponent(req) {
 /**
  * Can deploy a component in multiple ways. If a 'package' is provided all it will do is write that package to
  * harperdb-config, when HDB is restarted the package will be installed in hdb/node_modules. If a base64 encoded string is passed it
- * will write string to a tar file in the hdb/components dir, then it will extract the tar in hdb/components.
+ * will write string to a temp tar file and extract that file into the deployed project in hdb/components.
  * @param req
  * @returns {Promise<string>}
  */
@@ -395,48 +387,54 @@ async function deployComponent(req) {
 		throw new Error("'payload' or 'package' must be provided");
 	}
 
-	const path_to_project = path.join(cf_dir, project);
-	// check if the project exists, if it doesn't, create it.
-	await fs.ensureDir(path_to_project);
+	if (payload) {
+		const path_to_project = path.join(cf_dir, project);
+		pkg = 'file:' + path_to_project;
+		// check if the project exists, if it doesn't, create it.
+		await fs.ensureDir(path_to_project);
 
-	// Create a temp file to store project tar in. Check that is doesn't already exist, if it does create another path and test.
-	let temp_file_path;
-	let temp_file_exists;
-	do {
-		temp_file_path = path.join(TMP_PATH, uuidV4() + '.tar');
-		temp_file_exists = await fs.pathExists(temp_file_path);
-	} while (temp_file_exists);
+		// Create a temp file to store project tar in. Check that is doesn't already exist, if it does create another path and test.
+		let temp_file_path;
+		let temp_file_exists;
+		do {
+			temp_file_path = path.join(TMP_PATH, uuidV4() + '.tar');
+			temp_file_exists = await fs.pathExists(temp_file_path);
+		} while (temp_file_exists);
 
-	// pack the directory
-	await fs.outputFile(temp_file_path, payload, { encoding: 'base64' });
+		// pack the directory
+		await fs.outputFile(temp_file_path, payload, { encoding: 'base64' });
 
-	// extract the reconstituted file to the proper project directory
-	const stream = fs.createReadStream(temp_file_path);
-	stream.pipe(tar.extract(path_to_project));
-	await new Promise((resolve) => stream.on('end', resolve));
+		// extract the reconstituted file to the proper project directory
+		const stream = fs.createReadStream(temp_file_path);
+		stream.pipe(tar.extract(path_to_project));
+		await new Promise((resolve) => stream.on('end', resolve));
 
-	// delete the file
-	await fs.unlink(temp_file_path);
-
-	// HDB will package a component into a 'package` folder, this is done for npm install,
-	// however if we are here npm install is not being used, so we copy/remove component code up one dir.
-	const dir = await fs.readdir(path_to_project);
-	if (
-		(dir.length === 1 && dir.includes('package')) ||
-		(dir.length === 2 && dir.includes('package') && dir.includes('.DS_Store'))
-	) {
-		await fs.copy(path.join(path_to_project, 'package'), path_to_project);
-		await fs.remove(path.join(path_to_project, 'package'));
+		// delete the file
+		await fs.unlink(temp_file_path);
 	}
+
+	// Adds package to harperdb-config and then relies on restart to call install on the new app
+	config_utils.updateConfigValue(`${project}_package`, pkg, undefined, false, false, true);
 
 	return `Successfully deployed: ${project}`;
 }
-
 /**
  * Gets a JSON directory tree of the components dir and all nested files/folders
  * @returns {Promise<*>}
  */
-async function getComponentFiles() {
+async function getComponents() {
+	const all_config = config_utils.getConfiguration();
+	let comps = [];
+	for (const element in all_config) {
+		if (all_config[element]?.package) {
+			// Do not return packages that are file paths.
+			if (all_config[element].package.startsWith('file:')) {
+				continue;
+			}
+			comps.push(Object.assign(all_config[element], { name: element }));
+		}
+	}
+
 	// Recursive function that will traverse the components dir and build json
 	// directory tree as it goes.
 	const walk_dir = async (dir, result) => {
@@ -467,7 +465,7 @@ async function getComponentFiles() {
 
 	return walk_dir(env.get(terms.CONFIG_PARAMS.CUSTOMFUNCTIONS_ROOT), {
 		name: env.get(terms.CONFIG_PARAMS.CUSTOMFUNCTIONS_ROOT).split(path.sep).slice(-1).pop(),
-		entries: [],
+		entries: comps,
 	});
 }
 
@@ -509,9 +507,37 @@ async function setComponentFile(req) {
 
 	const options = req.encoding ? { encoding: req.encoding } : { encoding: 'utf8' };
 	const path_to_comp = path.join(env.get(terms.CONFIG_PARAMS.CUSTOMFUNCTIONS_ROOT), req.project, req.file);
-	await fs.ensureFile(path_to_comp);
-	await fs.outputFile(path_to_comp, req.payload, options);
+	if (req.payload !== undefined) {
+		await fs.ensureFile(path_to_comp);
+		await fs.outputFile(path_to_comp, req.payload, options);
+	} else {
+		await fs.ensureDir(path_to_comp);
+	}
+
 	return `Successfully set component: ` + req.file;
+}
+
+/**
+ * Deletes a component dir/file
+ * @param req
+ * @returns {Promise<string>}
+ */
+async function dropComponent(req) {
+	const validation = validator.dropComponentFileValidator(req);
+	if (validation) {
+		throw handleHDBError(validation, validation.message, HTTP_STATUS_CODES.BAD_REQUEST);
+	}
+
+	const project_path = req.file ? path.join(req.project, req.file) : req.project;
+	const path_to_comp = path.join(env.get(terms.CONFIG_PARAMS.CUSTOMFUNCTIONS_ROOT), project_path);
+
+	if (await fs.pathExists(path_to_comp)) {
+		await fs.remove(path_to_comp);
+	}
+
+	config_utils.deleteConfigFromFile([req.project]);
+
+	return 'Successfully dropped: ' + project_path;
 }
 
 module.exports = {
@@ -524,7 +550,8 @@ module.exports = {
 	dropCustomFunctionProject,
 	packageComponent,
 	deployComponent,
-	getComponentFiles,
+	getComponents,
 	getComponentFile,
 	setComponentFile,
+	dropComponent,
 };
